Move output mappings also int he wrong namespace
