+++ src/OSPSuite.Core/Serialization/Xml/OutputMappingsXmlSerializer.cs
@@ -1,4 +1,5 @@
-﻿using OSPSuite.Core.Domain.ParameterIdentifications;
+﻿using OSPSuite.Core.Domain;
+using OSPSuite.Core.Domain.ParameterIdentifications;
 
 namespace OSPSuite.Core.Serialization.Xml
 {
